requirements.txt
pandas>=1.3.0
numpy>=1.21.0
google-genai>=1.0.0
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=0.19.0
